add set method
Alright, it also refactors the client to use an aiohttp session for
connection pooling in an asynchronous fashion.  This allows us to reduce
the reliance on a custom fetch method and increase our use of aiohttp.

The number of test cases generated is probably extravagent but I don't
know how to prune it back (I'm not sure what good edges are).